LayoutTests/http/tests/security/window-name-after-cross-origin-aux-frame-navigation-expected.txt

CONSOLE MESSAGE: window.name should equal "test" after a cross-origin auxiliary frame navigation.
CONSOLE MESSAGE: window.name equals "test".
